+++ src/main/java/io/github/authservice/crowdfund/feature/useraddress/response/AddUserAddressResponse.java
@@ -2,6 +2,6 @@ package io.github.authservice.crowdfund.feature.useraddress.response;
 
 public record AddUserAddressResponse(
         String message,
-        Integer addressId
+        Long addressId
 ) {
 }
